# perf: add sharded/batch runners and dense model helpers