Summary

This PR introduces a high-performance inference pipeline for chess diagram → FEN conversion.
It adds batch and sharded batch runners for non-interactive processing, as well as a dense model path (single-shot 512×512 → (8×8×13)) for speedups.

Changes

bench/:

run_project_batch.py: sequential batch runner

run_project_sharded.py: parallel batch runner with configurable workers

fastpath_wrapper.py: dense vs. tile-based inference, benchmarking utilities

additional unit tests (test_dense.py, test_tiles.py, …)

speedkit/:

project_model_loader.py: refactored to support dense model loading

orient.py: board orientation heuristics

helpers for FEN generation and I/O

Benchmarks

Run on Mac (MPS, 3 workers):

Files: 37 | workers: 3 | total: 129.90s | ~3510.8 ms/item | shard fails: 0
→ log: bench/out_project_sharded.txt


This demonstrates a ~3.5s average per diagram, a significant improvement over the previous ~10s+.

Notes

Torch/vision device selection now prioritizes mps, then cuda, then cpu.

Dense model currently loads best_model_fen_0.943_2024-04-19-09-31-24.pth.

Still WIP: accuracy tuning and reducing misclassifications.