@frans860 2025-09-04
Hi Jost,

Thank you for your detailed feedback — and for taking the time to review my PR. I completely understand your hesitation about merging it.

You’re right: the comments ended up in Dutch, which is simply the language I think and learn in. I didn’t realize that it would make the code harder to review for others, but I’ll make sure to use English in the future.

I also want to mention that part of my motivation is that I’m still learning while coding, so this PR was also a way for me to explore new approaches and techniques. My main goal is to eventually process full PDFs with chess diagrams and convert them to FEN strings. That’s why speed and batch-processing are important in my use case.

I’ll also be sharing this project with someone else who runs Windows, so if you have any advice or best practices on how to set things up smoothly across platforms, it would be greatly appreciated. Given your experience, I’d really value your perspective.

Thanks again for your honest response — and for maintaining this project in the first place!

Best,
Frans